weather-collector/requirements.txt
pika==1.3.2
requests==2.31.0
python-dotenv==1.0.0

